pyproject.toml
[project]
name = "langgraph_eng_package"
version = "0.0.1"
dependencies = [
  "langgraph",
  "langchain_anthropic",
  "langchain_core",
  "langchain_openai",
  "gitpython",
  "python-dotenv",
  "pydantic",
  "aider-chat",
  "semantic-router"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"